fix: vercel 수정, 파일 위치 변경
