+++ src/components/feature/Modal/CheckModal/CheckModal.module.scss
